requirements.txt
alembic==1.7.5
attrs==21.2.0
Babel==2.9.1
blinker==1.4
boto3==1.20.41
botocore==1.23.41
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
coverage==6.2
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.3
Flask==2.0.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Scss==0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.1.6
MarkupSafe==2.0.1
notifications-python-client==6.3.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pyScss==1.3.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.29
toml==0.10.2
urllib3==1.26.8
waitress==2.0.0
Werkzeug==2.0.2
WTForms==3.0.0

